+++ src/app/buses/page.tsx
@@ -1,11 +1,12 @@
+import { BackBtn } from "@/backBtn";
+import Header from "@/header";
+import { BusPageListSkeleton } from "@/navlist/busNavList";
+import ScrollToTopButton from "@/scrollToTopBtn";
+import { BusTag } from "@/tags";
 import Link from "next/link";
 import { Suspense } from "react";
 import { IoChevronBackSharp, IoChevronForwardSharp } from "react-icons/io5";
 import { api } from "../../trpc/server";
-import { BackBtn } from "../_components/backBtn";
-import Header from "../_components/header";
-import ScrollToTopButton from "../_components/scrollToTopBtn";
-import { BusTag } from "../_components/tags";
 
 export default async function Page() {
   return (
@@ -20,8 +21,8 @@ export default async function Page() {
             </BackBtn>
           </Suspense>
         </div>
-        <Suspense fallback={<BusListSkeleton />}>
-          <BusList />
+        <Suspense fallback={<BusPageListSkeleton />}>
+          <BusPageList />
         </Suspense>
       </div>
       <ScrollToTopButton />
@@ -29,7 +30,7 @@ export default async function Page() {
   );
 }
 
-async function BusList() {
+async function BusPageList() {
   const buses = await api.bus.getAll.query();
 
   return (
@@ -50,28 +51,6 @@ async function BusList() {
           </div>
           <IoChevronForwardSharp size={24} />
         </Link>
-      ))}
-    </>
-  );
-}
-
-export function BusListSkeleton() {
-  return (
-    <>
-      {[...Array(6).keys()].map((count) => (
-        <div
-          key={count}
-          className=" flex w-full flex-row items-center gap-3 border-t-2 border-white p-3 text-white transition-all hover:bg-slate-600"
-        >
-          <div className=" flex flex-1 flex-col gap-2">
-            <div className=" flex w-full flex-row items-center gap-2">
-              <div className=" aspect-square w-10 animate-pulse rounded-md bg-slate-400" />
-              <div className=" h-5 w-full animate-pulse overflow-clip text-ellipsis text-nowrap rounded-md bg-slate-400 text-lg font-bold animation-delay-150" />
-            </div>
-            <div className=" h-5 w-8/12 animate-pulse rounded-md bg-slate-400"></div>
-          </div>
-          <IoChevronForwardSharp size={24} />
-        </div>
       ))}
     </>
   );
