+++ src/app/stops/loading.tsx
@@ -1,9 +1,9 @@
+import { BackBtn } from "@/backBtn";
+import Header from "@/header";
+import { StopListSkeleton } from "@/navlist/stopNavList";
+import ScrollToTopButton from "@/scrollToTopBtn";
 import { Suspense } from "react";
 import { IoChevronBackSharp } from "react-icons/io5";
-import { BackBtn } from "../_components/backBtn";
-import Header from "../_components/header";
-import ScrollToTopButton from "../_components/scrollToTopBtn";
-import { StopListSkeleton } from "./page";
 
 export default async function Page() {
   return (
